+++ src/webpack/common/components.ts
@@ -16,12 +16,11 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-import { LazyComponent } from "@utils/misc";
 import type Components from "discord-types/components";
 import type { ComponentType } from "react";
 
 // eslint-disable-next-line path-alias/no-relative
-import { filters, findByCode, findByCodeLazy, findByPropsLazy } from "../webpack";
+import { filters, findByPropsLazy } from "../webpack";
 import { makeWaitForComponent } from "./internal";
 import type * as t from "./types/components";
 
@@ -41,17 +40,16 @@ export const Button = makeWaitForComponent("Button", ["Hovers", "Looks", "Sizes"
 export const Switch = makeWaitForComponent("Switch", filters.byCode("tooltipNote", "ringTarget"));
 export const Tooltip = makeWaitForComponent<Components.Tooltip>("Tooltip", ["Positions", "Colors"]);
 export const Timestamp = makeWaitForComponent("Timestamp", filters.byCode(".Messages.MESSAGE_EDITED_TIMESTAMP_A11Y_LABEL.format"));
+export const TextArea = makeWaitForComponent<React.ComponentType<React.PropsWithRef<any>>>("TextArea", ["handleSetRef", "textArea"]);
 export const TextInput = makeWaitForComponent<Components.TextInput>("TextInput", ["defaultProps", "Sizes", "contextType"]);
 export const Text = makeWaitForComponent<ComponentType<t.TextProps>>("Text", m => {
     if (typeof m !== "function") return false;
     const s = m.toString();
     return (s.length < 1500 && s.includes("data-text-variant") && s.includes("always-white"));
 });
+export const Select = makeWaitForComponent("Select", ["optionClassName", "popoutPosition", "autoFocus", "maxVisibleItems"]);
+export const Slider = makeWaitForComponent("Slider", ["closestMarkerIndex", "stickToMarkers"]);
 
-export const TextArea = findByCodeLazy("handleSetRef", "textArea") as React.ComponentType<React.PropsWithRef<any>>;
 export const ButtonWrapperClasses = findByPropsLazy("buttonWrapper", "buttonContent") as Record<string, string>;
-export const Select = LazyComponent(() => findByCode("optionClassName", "popoutPosition", "autoFocus", "maxVisibleItems"));
-export const Slider = LazyComponent(() => findByCode("closestMarkerIndex", "stickToMarkers"));
-
 export const Margins: t.Margins = findByPropsLazy("marginTop20");
-export const ButtonLooks = findByPropsLazy("BLANK", "FILLED", "INVERTED") as Record<"FILLED" | "INVERTED" | "OUTLINED" | "LINK" | "BLANK", string>;
+export const ButtonLooks: t.ButtonLooks = findByPropsLazy("BLANK", "FILLED", "INVERTED");
