+++ src/webpack/common/index.ts
@@ -21,5 +21,6 @@ export * from "./menu";
 export * from "./react";
 export * from "./stores";
 export * as ComponentTypes from "./types/components.d";
+export * as MenuTypes from "./types/menu.d";
 export * from "./utils";
